+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## V1.0.7
+
+Use TPNS iOS SDK 1.2.9.0
+Use TPNS Android SDK 1.2.2.4
+
+## V1.0.6
+
+* Fix api setAccount() not compatible with old version flutter binder
+* Add Api getOtherPushType for Android
+
 ## V1.0.5
 
 * Add Api getOtherPushToken for Android